Fix ZoomSlider thumb position after click in slider
Clicking in the zoomslider container to set the position without dragging placed the
thumb at the wrong position. The padding left/top of the container needs to be added to
the event position and by adding 0.5 the clicked pixel is converteded to the starting
edge (top/left)

This now correctly handles a border thickness of the ZoomSlider container.