Drag map at pointer down position
The map should be panned at the pointer down position, not the first position
where the moved distance is greater than the move tolerance.

This also fixes a small bug where the very first drag event handled by this
interaction will not update the kinetic because the lastPointerCount_ variable
was not updated on pointer down. Though it is never reset on pointer up.